requirements.txt
tensorflow==2.4.1
tensorflow-datasets==4.2.0
tensorflow-probability==0.12.1
pyYAML
numpy
scikit-image==0.18.1
imageio
minerl-navigate
